mm/memblock: implement memblock_advise_size_order and probe functions
Hotplug memory sources may have opinions on what the memblock size
should be - usually for alignment purposes.  For example, CXL memory
extents as small as 256MB with a matching alignment in physical memory.

Implement memblock_advise_size_order for use during early init, prior
to allocator and smp init, for software to advise the system as to what
the preferred block size should be.

The probe function is meant for arch_init code to fetch this value
once during memblock size calculation. Use of the advisement value
is arch-specific, and no guarantee is made that it will be used.

Calls to either function after probe results in -EBUSY to signal that
advisement is ignored or that memblock_get_size_bytes should be used.

Suggested-by: Ira Weiny <ira.weiny@intel.com>
Signed-off-by: Gregory Price <gourry@gourry.net>