vmscan,migrate: fix double-decrement on node stats when demoting pages
When numa balancing is enabled with demotion, vmscan will call
migrate_pages when shrinking LRUs.  Successful demotions will
cause node vmstat numbers to double-decrement, leading to an
imbalanced page count.  The result is dmesg output like such:

$ cat /proc/sys/vm/stat_refresh

[77383.088417] vmstat_refresh: nr_isolated_anon -103212
[77383.088417] vmstat_refresh: nr_isolated_file -899642

This negative value may impact compaction and reclaim throttling.

The double-decrement occurs in the migrate_pages path:

caller to shrink_folio_list decrements the count
  shrink_folio_list
    demote_folio_list
      migrate_pages
        migrate_pages_batch
          migrate_folio_move
            migrate_folio_done
              mod_node_page_state(-ve) <- second decrement

This path happens for SUCCESSFUL migrations, not failures. Typically
callers to migrate_pages are required to handle putback/accounting for
failures, but this is already handled in the shrink code.

When accounting for migrations, instead do not decrement the count
when the migration reason is MR_DEMOTION. As of v6.11, this demotion
logic is the only source of MR_DEMOTION.

Signed-off-by: Gregory Price <gourry@gourry.net>